pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "narrativecluster"
version = "0.1.0"
description = "Claim-level narrative clustering with HNSW + Leiden + neighbor-vote assignment"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Patrick Gerard" }
]
keywords = ["narrative", "clustering", "FAISS", "Leiden", "NLP", "discourse"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "faiss-cpu>=1.7",
    "python-igraph>=0.11",
    "leidenalg>=0.10",
    "tqdm>=4.65",
    "pyarrow>=12.0",
]

[project.optional-dependencies]
gpu = ["faiss-gpu>=1.7"]
dev = ["pytest>=7.0", "pytest-cov", "black", "ruff"]

[project.scripts]
narrativecluster = "narrativecluster.cli:main"

[project.urls]
Homepage = "https://github.com/patrickgerard/narrativecluster"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
